[ias-verify] move verify cert to main
